添加背景视频 一些弹设 河童道中符卡

+++ src/game/bullet_pool.hpp
@@ -39,7 +39,7 @@ public:
         godot::Vector2 position;
         /// @brief 每帧速度
         godot::Vector2 velocity;
-        float rotation = 0.0f;
+        double rotation = 0.0f;
         /// @brief 存活帧数
         int lifetime = 0; 
         bool active = false;
